REPO-1530 [PATCH] REPO-182: store mixintypes also in persistent store
The following patches are a fix for JCR-2633.
This changes the on disk format (forward and backward compatible) to
also store the jcr:mixinTypes property.  By storing these properties,
also the modification count is persisted, which is needed for a check
whether only the mixin types have changes, and the original node was
purged from cache.