CMS-14602 [PATCH] REPO-191: Parameterized path entries in Jackrabbit
The following patches implement the support for parameterized paths.
- The patch to BatchedItemOperations is a removal of a piece of
code that performs a sanity check in case of sharable nodes.  Since we
do not use this, and this check (but the check alone) conflicts with
parameterized (free text) searches, the check is disabled.
- The CargoPathElement then introduces a new kind of path element.
A path normally is contained of path elements, in a simple path
each path element is a name part in a relative path, i.e. /a/b/c
contains name elements a, b and name element c.  Other path elements
are things like a identifier elements, such as the root node which is
named /jcr:root, or a parent path element, a current path element,
etcetera.  A CargoPathElement indicates that it is both a named
elements (denotesName() returns true) as an identifier element.  This
combination does not occur in JackRabbit, such that later on this can
be used to identify the element by.  The sole purpose of a
CargoPathElement is to carry additional payload (cargo) that is user-
defined, and otherwise it acts like a normal name element.
- The patch to PathBuilder forces a lookup from the root (which is
done a.o when you look up a node using Session.getNode(String)), to
be resolved using the normal path factory (the one that can be
overridden), rather than a hardcoded entry set at the root node.
- Within the PathFactoryImpl there is one patch that creates
a CargoPathElement when it encounters a '[' in the name.  The
pathbuilder would normally have transformed entries like a[2]
into a name element with an additional parameter with value 2, however
you can override the pathbuilder to pass certain other information
a[b] and let this be retained in the name.
- All other patches, including the other patch in PathFactoryImpl are
all just about preserving a CargoPathElement when building one path
from another path.

(cherry picked from commit 39bf4045bf08df76097d553fb36d69c1129892c9)