REPO-1433 Support for callbacks synchronizing on either updateMonitor or luceneExportMonitor
When preparing a Lucene Export, we need to have a lock on the updateMonitor to make sure indexes
do not change while calculating how many bytes we need to copy per index.

After this preparing phase, we want to do the actual export of Lucene indexes. However during this
phase, we do not want the org.apache.jackrabbit.core.query.lucene.MultiIndex#replaceIndexes to kick
in as this method can add 'DeleteIndex' actions to the executor which as a result can close an index
while we still need it, resulting in a org.apache.lucene.store.AlreadyClosedException