REPO-1750 Avoid concurrent modification exceptions
See the patch in https://issues.apache.org/jira/browse/JCR-3471
but note that the patch in JCR is not the complete story: It only
reduces the odds of the concurrent modification exceptions.

If you use

for (Name versionName : new HashSet<Name>(nameCache.keySet())) {

the result is that a new HashSet is created by iterating through the
nameCache its keys. However, this iterator is still bound to the
original nameCache map. As a result, new
HashSet<Name>(nameCache.keySet())) can still result in the
ConcurrentModificationException.

The most straightforward fix is making the fixLegacy method
synchronized, however, I do not know all the bits and bytes in this
code area and don't want to accidental cause a deadlock. Hence, I will
rewrite it to:

private void fixLegacy() throws RepositoryException {
    if (rootVersion.getSuccessors().isEmpty()) {
        HashSet<Name> names;
        synchronized (this) {
            names = new HashSet<>(nameCache.keySet());;
        }
        for (Name versionName : names) {
            InternalVersionImpl v = createVersionInstance(versionName);
            v.legacyResolveSuccessors();
        }
    }
}

I've confirmed that 'nameCache' is a private field and all
modifications to it are synchronized on the InternalVersionHistoryImpl
instance.