REPO-2026 Ignore index-rule for not (yet) registered nodetypes
This is a fix/workaround for the changes introduced by https://issues.apache.org/jira/browse/JCR-4093 which now assume/require
the nodetype specified to be already registered in the NodeTypeRegistry.
That assumption only can be true for the standard/pre-defined nodetypes (with jcr:, mix: etc. prefixes), and causes a chicken-egg problem
for custom nodetypes. Luckily the IndexingConfiguration also is registered as listener for changes in the NodeTypeRegistry, which means that
NoSuchNodeTypeExceptions can safely be ignored and will 'self-repair' as soon as such nodetypes are registered later during the startup.