REPO-2268 Correct the synchronization on the 'updateMonitor/updateInProgress'
While the updateInProgress = true it means something is happening with
the index, for example a document  is being added/deleted, etc. So, as
long as  updateInProgress = true, we need to wait. After updateInProgress
has become false, we can try to invoke the 'Callable" synchronized on
the 'updateMonitor' *only* if directly after the 'updateMonitor' has
been notified not another thread has already set 'updateInProgress' to
true again. Therefor, we need another 'updateInProgress' check, and if
it is true, we still cannot proceed the call to the Callable but instead,
have to wait again.

Theoretically there is room for a stack overflow but that is just
*extremely* unlikely so can be safely ignored