CMS-14788 Cache the binary lenght in InternalValue instance
InternalValue instances are part of the cached bundle objects
representing a jcr node.

This means that whenever the node changes, the InternalValue instances
are also replaced.

This means, that we can cache a result in an InternalValue
instance variable without problem. This is super important for
the binary length since this one always falls back to a database
call requesting the meta info of the binary in the datastore

This simple fix avoids the expensive db call. When the resource
changes, the InternalValue instance is also replaced, and as a
result, the length of the binary also gets correctly uddated: it
just piggybacks on the bundle cache logic handling invalidation