ENT-7352 Ordered Synchronous event listeners to avoid hiccups
Jackrabbit has two kind of event listeners:

# Synchronous Event Listeners
# Asynchronous Event Listeners

The synchronous event listeners are executed blocking the request
which triggers a transaction. The asynchronous listeners are
updated later asynchronously by a different thread than the
thread making the transaction.

It is tricky that JR has the Synchronous Event Listeners as a Set
resulting in a non deterministic order. However the first
registered synchronous event listener, the SearchManager, is by
far the most expensive in general as it does the Lucene indexing
which can take considerable time. It is better that other
synchronous event listeners are notified AFTER the SearchManager:
An event listener is typically used to invalidate some model. You
do not want to block reloading such a model for longer than
necessary, however, if the model is invalidated and after that the
expensive SearchManager listener is notified, the model loading can
be blocked for a long time. For this reason, keep a List instead
of a Set. This works, as the SearchManager is always the first
listener to be registered